refactor: 날짜 포맷팅 유틸 함수 통합으로 중복 코드 제거
