+++ components/home/WeeklyCard.tsx
@@ -1,10 +1,10 @@
 import { useState } from "react";
 
-import dayjs from "dayjs";
 import { useRouter } from "expo-router";
 import { Image, Pressable, Text, View } from "react-native";
 
 import { formatAddress } from "@/utils/addressUtils";
+import { formatShortDate } from "@/utils/dateUtils";
 
 interface WeeklyContentItem {
   contentId: number;
@@ -20,8 +20,6 @@ export default function WeeklyCard({ item }: { item: WeeklyContentItem }) {
   const [imageLoaded, setImageLoaded] = useState(false);
 
   const handlePress = () => router.push(`/detail/${item.contentId}`);
-
-  const formatDate = (date: string) => dayjs(date).format("YY.MM.DD");
 
   const hasImage = item.image && item.image.trim() !== "";
   const imageSource = hasImage
@@ -70,7 +68,7 @@ export default function WeeklyCard({ item }: { item: WeeklyContentItem }) {
           {formatAddress(item.address)}
         </Text>
         <Text className="text-sm font-normal text-[#707070]">
-          {formatDate(item.startDate)} ~ {formatDate(item.endDate)}
+          {formatShortDate(item.startDate)} ~ {formatShortDate(item.endDate)}
         </Text>
       </View>
     </Pressable>
